+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - dirconfig: add subscription setting with comment [PR #2468]
 - hyper-v: fix restores with portable data [PR #2469]
 
+### Fixed
+- VMware Plugin: Fix NVRAM backup when datacenter is not in root folder [PR #2461]
+
 ## [25.0.0] - 2025-12-01
 
 ### Breaking changes
@@ -2171,6 +2174,7 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2452]: https://github.com/bareos/bareos/pull/2452
 [PR #2456]: https://github.com/bareos/bareos/pull/2456
 [PR #2458]: https://github.com/bareos/bareos/pull/2458
+[PR #2461]: https://github.com/bareos/bareos/pull/2461
 [PR #2468]: https://github.com/bareos/bareos/pull/2468
 [PR #2469]: https://github.com/bareos/bareos/pull/2469
 [unreleased]: https://github.com/bareos/bareos/tree/master